fix(cluster,native): minor correctness fixes in applier and plan builders
- `metadata_group/applier.rs`: remove stale doc comment that described
  `apply_topology_change` as backward-compatible with old test wiring;
  the field is simply optional.
- `plan_builder/graph.rs`: correct `clamped_depth` parameter from
  `Option<u64>` to `Option<u32>` to match the wire field type.
- `plan_builder/vector.rs`: add explicit `u32` bounds check on
  `vector_id` with a structured `BadRequest` error rather than silently
  truncating the surrogate.

+++ nodedb/src/control/server/native/dispatch/plan_builder/graph.rs
@@ -14,7 +14,7 @@ use super::parse_direction;
 /// rejecting out-of-range values rather than forwarding them to the
 /// engine. Mirrors the pgwire ingress so no entry point can saturate
 /// traversal with an unbounded fan-out.
-fn clamped_depth(value: Option<u64>, default: usize, field: &str) -> crate::Result<usize> {
+fn clamped_depth(value: Option<u32>, default: usize, field: &str) -> crate::Result<usize> {
     let v = value.map(|v| v as usize).unwrap_or(default);
     if v > MAX_GRAPH_TRAVERSAL_DEPTH {
         return Err(crate::Error::BadRequest {
